Change: Comms interfaces outline
