--- src/app/comms/types.ts
@@ -1,9 +1,0 @@
-/*
-	File: src/shared/comms/types.ts
-	cpuabuse.com
-*/
-
-/**
- * Communication types to literally use to send physical data.
- * "X" stands for exchange.
- */
