+++ src/app/server/things/exclusive/immovable.ts
