+++ src/app/server/things/exclusive/immovable/wall.ts
