+++ src/app/server/things/stackable/items/gold.ts
