+++ src/app/shared/comms/uuid.ts
@@ -1,0 +1,48 @@
+/**
+ * Uuid interfaces.
+ */
+
+/**
+ * Uuid alias for [[CommsMap]].
+ */
+export type CommsMapUuid = Uuid;
+
+/**
+ * Uuid alias for [[Instance]].
+ */
+export type InstanceUuid = Uuid;
+
+/**
+ * Uuid alias for [[Kind]].
+ */
+export type KindUuid = Uuid;
+
+/**
+ * Uuid alias for [[Location]].
+ */
+export type LocationUuid = Uuid;
+
+/**
+ * Uuid alias for [[Mode]].
+ */
+export type ModeUuid = Uuid;
+
+/**
+ * Uuid alias for [[Occupant]].
+ */
+export type OccupantUuid = Uuid;
+
+/**
+ * Uuid alias for [[View]].
+ */
+export type ViewUuid = Uuid;
+
+/**
+ * Uuid alias for [[World]].
+ */
+export type WorldUuid = Uuid;
+
+/**
+ * Alias for uuid.
+ */
+export type Uuid = string;
